DataAnalysis uptade - now its more readable and has DropDownMenu to choose witch datas chart sould represent.

+++ app/src/main/java/com/example/apogemixconnect/model/Data/FlightDataDB/FlightDatas.kt
@@ -14,7 +14,7 @@ data class FlightDatas(
     val gpsLat: Float,
     val gpsLng: Float,
     val gpsAlt: Float,
-    val time: Int,
+    val time: Float,
     val temperature: Float,
     val pressure: Float,
     val height: Float,
